dom/media/test/test_mkv_playback.html

<!DOCTYPE html>
<html>
<head>
    <title>MKV playback test</title>
    <link rel="stylesheet" href="/tests/SimpleTest/test.css">
    <script src="/tests/SimpleTest/SimpleTest.js"></script>
    <script type="text/javascript" src="manifest.js"></script>
</head>
<body>
<pre id="test">
<script>
/**
 * This test verifies supported codecs in MKV files play to the end.
 */
add_task(async function test_supported_mkv_sources() {
  await SpecialPowers.pushPrefEnv({
    set: [["media.mkv.enabled", true]]
  });
  for (const test of gMKVtests) {
    let v = document.createElement("video");
    document.body.appendChild(v);
    v.src = test.name;
    v.preload = "auto";
    info(`Starting to play ${test.name}`);
    let endPromise = once(v, "ended");
    ok(await v.play().then(_ => true, _ => false), "media failed to play");
    await endPromise;
    ok(`${test.name} played to the end`);
    removeNodeAndSource(v);
  }
});
</script>
</pre>
</body>
</html>
